BSPM2021T3-95 Splited the backend functionality into bluesprint files
